Make limit check and additions an atomic operation
Due to the fact that checking whether the limit is exceeded and adding a new item to the list are different operations, if two sidekiq jobs are performed at the same time and check for the limit at the same time, both can be returned true, and proceed with executing the work.

The addition of a semaphore for the check & add to limit operations ensures that it's only done by one job at the time, and fixes the race condition.