Update semaphore redis to initialize a new conn
The redis-semaphore gem advises [against reusing the redis connection across threads](https://github.com/dv/redis-semaphore/issues/18) because it causes a Deadlock. We actually experience this and the solution was to initialize a new redis connection when checking for the semaphore.

I additionally added a staleness check to avoid locking everything forever should the semaphore hang for some other reason.